+++ delete_test.go
@@ -22,17 +22,18 @@ func TestDelete(t *testing.T) {
 		fmt.Printf("*****db-connection-error: %v\n", err.Error())
 		return
 	}
-	audit := Audit{}
+	model := Audit{}
+	modelPtr := AuditPtr{}
 	crudParams := CrudParamsType{
 		AppDb:        dbc,
-		ModelRef:     audit,
-		ModelPointer: &audit,
+		ModelRef:     model,
+		ModelPointer: &modelPtr,
 		TableName:    DeleteTable,
 		UserInfo:     TestUserInfo,
 		RecordIds:    []string{},
 		QueryParams:  QueryParamType{},
 	}
-	var crud = NewCrud(crudParams, CrudParamOptions)
+	crud := NewCrud(crudParams, CrudParamOptions)
 
 	mctest.McTest(mctest.OptionValue{
 		Name: "should prevent the delete of all table records and return removeError:",
@@ -45,40 +46,43 @@ func TestDelete(t *testing.T) {
 	})
 
 	mctest.McTest(mctest.OptionValue{
-		Name: "should delete record by Id and return success[delete-record-method]:",
+		Name: "should delete record by Id and return success or notFound[delete-record-method]:",
 		TestFunc: func() {
 			crud.TableName = DeleteTable
 			crud.RecordIds = []string{DeleteAuditById}
 			crud.QueryParams = QueryParamType{}
 			// get-record method params
 			res := crud.DeleteRecord()
-			fmt.Printf("delete-all: %#v \n", res)
-			mctest.AssertEquals(t, res.Code, "success", "delete-by-id should return code: success")
+			fmt.Printf("delete-by-id-result: %#v \n", res)
+			resCode := res.Code == "success" || res.Code == "notFound"
+			mctest.AssertEquals(t, resCode, true, "delete-by-id should return code: success or notFound")
 		},
 	})
 
 	mctest.McTest(mctest.OptionValue{
-		Name: "should delete records by Ids and return success[delete-record-method]:",
+		Name: "should delete records by Ids and return success or notFound[delete-record-method]:",
 		TestFunc: func() {
 			crud.TableName = DeleteTable
 			crud.RecordIds = DeleteAuditByIds
 			crud.QueryParams = QueryParamType{}
 			// get-record method params
 			res := crud.DeleteRecord()
-			fmt.Printf("delete-all: %#v \n", res)
-			mctest.AssertEquals(t, res.Code, "success", "delete-by-id should return code: success")
+			fmt.Printf("delete-by-ids-result: %#v \n", res)
+			resCode := res.Code == "success" || res.Code == "notFound"
+			mctest.AssertEquals(t, resCode, true, "delete-by-ids should return code: success or notFound")
 		},
 	})
 
 	mctest.McTest(mctest.OptionValue{
-		Name: "should delete records by query-params and return success[delete-record-method]:",
+		Name: "should delete records by query-params and return success or notFound[delete-record-method]:",
 		TestFunc: func() {
 			crud.TableName = DeleteTable
 			crud.RecordIds = []string{}
 			crud.QueryParams = DeleteAuditByParams
 			res := crud.DeleteRecord()
-			fmt.Printf("delete-all: %#v \n", res)
-			mctest.AssertEquals(t, res.Code, "success", "delete-by-params-log should return code: success")
+			fmt.Printf("delete-by-params-result: %#v \n", res)
+			resCode := res.Code == "success" || res.Code == "notFound"
+			mctest.AssertEquals(t, resCode, true, "delete-by-params should return code: success or notFound")
 		},
 	})