+++ HISTORY.md
@@ -2,5 +2,5 @@
 
 ## Version 0.1.0
 
-- Release date: Oct 23, 2021
-- baseline-version-for-mccrud
+- Release date: April 29, 2022
+- baseline-version-for mcdbcrud package